hitas/views: move paginator under 'hitas/views/paginator'
